requirements.txt
Flask==2.2.5               # Stable version compatible with Werkzeug
Werkzeug==2.2.3             # Compatible with Flask 2.2.5
Flask-Cors==3.0.10
google-generativeai==0.3.0
numpy==1.24.2
pandas==1.5.3
matplotlib==3.7.1
tensorflow==2.14.0
gunicorn==21.2.0
python-dotenv==1.0.0
